Improve overall UX for gameplay

+++ Assets/JANOARG.Client/Scripts/Behaviors/Player/PlayerScreenResultDetails.cs
@@ -182,7 +182,7 @@ namespace JANOARG.Client.Behaviors.Player
                         new Vector2(
                             item.Time / PlayerScreen.sTargetSong.Clip.length,
                             item.Type == HitObjectHistoryType.Timing
-                                ? Mathf.Clamp01(.5f - item.Offset / PlayerScreen.sMain.GoodWindow / 2)
+                                ? Mathf.Clamp01((float)(.5f - item.Offset / PlayerScreen.sMain.GoodWindow / 2))
                                 : 0.5f
                         );
